api/views.py
from rest_framework import viewsets
from api.models import Anime
from api.serializers import AnimeSerializer

class AnimeViewSet(viewsets.ModelViewSet):
    queryset = Anime.objects.all()
    serializer_class = AnimeSerializer


