requirements.txt
asgiref==3.8.1
Django==5.0
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
gunicorn==23.0.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==24.1
sqlparse==0.5.1
tzdata==2024.1